refactor(ekb): move pipeline logic to app/ folder (Part A only)
