+++ pipelines/enterprise_knowledge_base/app/document_classification/bq_service/schemas.py
