+++ pipelines/enterprise_knowledge_base/app/document_classification/gcs_service/schemas.py
